+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_last/prefer_last_rule.dart
@@ -4,13 +4,13 @@ import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/token.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
 
+import '../../../../../utils/dart_types_utils.dart';
 import '../../../../../utils/node_utils.dart';
 import '../../../lint_utils.dart';
 import '../../../models/internal_resolved_unit_result.dart';
 import '../../../models/issue.dart';
 import '../../../models/replacement.dart';
 import '../../../models/severity.dart';
-import '../../dart_rule_utils.dart';
 import '../../models/common_rule.dart';
 import '../../rule_utils.dart';