+++ lib/src/analyzers/unused_code_analyzer/used_code_visitor.dart
@@ -4,6 +4,7 @@ import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/dart/element/element.dart';
 
+import '../../utils/flutter_types_utils.dart';
 import 'models/file_elements_usage.dart';
 import 'models/prefix_element_usage.dart';
 
@@ -154,6 +155,11 @@ class UsedCodeVisitor extends RecursiveAstVisitor<void> {
       return;
     }
 
+    // Usage in State<WidgetClassName> is not a sing of usage.
+    if (_isUsedAsNamedTypeForWidgetState(identifier)) {
+      return;
+    }
+
     // Record `importPrefix.identifier` into 'prefixMap'.
     if (_recordPrefixMap(identifier, element)) {
       return;
@@ -209,4 +215,11 @@ class UsedCodeVisitor extends RecursiveAstVisitor<void> {
     SimpleIdentifier identifier,
   ) =>
       target is VariableDeclaration && target.initializer == identifier;
+
+  bool _isUsedAsNamedTypeForWidgetState(SimpleIdentifier identifier) {
+    final grandGrandParent = identifier.parent?.parent?.parent;
+
+    return grandGrandParent is NamedType &&
+        isWidgetStateOrSubclass(grandGrandParent.type);
+  }
 }